+++ test/Elastica/Aggregation/TopHitsTest.php
@@ -20,8 +20,14 @@ class TopHitsTest extends BaseAggregationTest
         $mapping = new Mapping($index->getType('test'), [
             'tags' => ['type' => 'keyword'],
             'title' => ['type' => 'keyword'],
+            'my_join_field' => [
+                'type' => 'join',
+                'relations' => [
+                    'question' => 'answer'
+                ]
+            ]
         ]);
-        $index->getType('test')->setMapping($mapping);
+        $index->getType('questions')->setMapping($mapping);
 
         $index->getType('questions')->addDocuments([
             new Document(1, [
@@ -202,7 +208,6 @@ class TopHitsTest extends BaseAggregationTest
      */
     public function testAggregateUpdatedRecently()
     {
-        $this->markTestSkipped('ES6 update: the final mapping would have more than 1 type');
         $aggr = new TopHits('top_tag_hits');
         $aggr->setSize(1);
         $aggr->setSort(['last_activity_date' => ['order' => 'desc']]);
@@ -222,7 +227,6 @@ class TopHitsTest extends BaseAggregationTest
      */
     public function testAggregateUpdatedFarAgo()
     {
-        $this->markTestSkipped('ES6 update: the final mapping would have more than 1 type');
         $aggr = new TopHits('top_tag_hits');
         $aggr->setSize(1);
         $aggr->setSort(['last_activity_date' => ['order' => 'asc']]);
@@ -242,7 +246,6 @@ class TopHitsTest extends BaseAggregationTest
      */
     public function testAggregateTwoDocumentPerTag()
     {
-        $this->markTestSkipped('ES6 update: the final mapping would have more than 1 type');
         $aggr = new TopHits('top_tag_hits');
         $aggr->setSize(2);
 
@@ -261,7 +264,6 @@ class TopHitsTest extends BaseAggregationTest
      */
     public function testAggregateTwoDocumentPerTagWithOffset()
     {
-        $this->markTestSkipped('ES6 update: the final mapping would have more than 1 type');
         $aggr = new TopHits('top_tag_hits');
         $aggr->setSize(2);
         $aggr->setFrom(1);
@@ -282,7 +284,6 @@ class TopHitsTest extends BaseAggregationTest
      */
     public function testAggregateWithLimitedSource()
     {
-        $this->markTestSkipped('ES6 update: the final mapping would have more than 1 type');
         $aggr = new TopHits('top_tag_hits');
         $aggr->setSource(['title']);
 
@@ -302,7 +303,6 @@ class TopHitsTest extends BaseAggregationTest
      */
     public function testAggregateWithVersion()
     {
-        $this->markTestSkipped('ES6 update: the final mapping would have more than 1 type');
         $aggr = new TopHits('top_tag_hits');
         $aggr->setVersion(true);
 
@@ -320,7 +320,6 @@ class TopHitsTest extends BaseAggregationTest
      */
     public function testAggregateWithExplain()
     {
-        $this->markTestSkipped('ES6 update: the final mapping would have more than 1 type');
         $aggr = new TopHits('top_tag_hits');
         $aggr->setExplain(true);
 
@@ -338,7 +337,6 @@ class TopHitsTest extends BaseAggregationTest
      */
     public function testAggregateWithScriptFields()
     {
-        $this->markTestSkipped('ES6 update: the final mapping would have more than 1 type');
         $aggr = new TopHits('top_tag_hits');
         $aggr->setSize(1);
         $aggr->setScriptFields(['three' => new Script('1 + 2')]);
@@ -359,7 +357,6 @@ class TopHitsTest extends BaseAggregationTest
      */
     public function testAggregateWithHighlight()
     {
-        $this->markTestSkipped('ES6 update: the final mapping would have more than 1 type');
         $queryString = new SimpleQueryString('linux', ['title']);
 
         $aggr = new TopHits('top_tag_hits');
@@ -382,7 +379,6 @@ class TopHitsTest extends BaseAggregationTest
      */
     public function testAggregateWithFieldData()
     {
-        $this->markTestSkipped('ES6 update: the final mapping would have more than 1 type');
         $aggr = new TopHits('top_tag_hits');
         $aggr->setFieldDataFields(['title']);